Python/MyTimeZones/pyproject.toml
[project]
name = "mytimezones"
version = "1.0.0"
description = "A timezone management application built with Python and Tkinter"
requires-python = ">=3.8.1"
dependencies = [
    "pytz>=2023.3",
    "python-dateutil>=2.8.2",
    "ttkthemes>=3.2.2",
    "ttkwidgets>=0.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
# UV specific configurations
python-preference = "managed"
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]
